# xt_server_openApi_autogenerated.json
{
  "swagger": "2.0",
  "info": {
    "title": "XTVision server",
    "description": "All XTVsion server apis",
    "version": "1.0.0"
  },
  "host": "localhost:5001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "securityDefinitions": {},
  "paths": {
...
    "/notifMessages/": {
      "post": {
        "tags": [
          "NOTIF MESSAGES"
        ],
        "summary": "save a Notif Predefined Message",
        "description": "",
        "parameters": [
          {
            "name": "message",
            "in": "body",
            "description": "save notif message ",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "string"
                },
                "body": {
                  "type": "string",
                  "example": "string"
                },
                "messageTtl": {
                  "type": "number",
                  "example": 3600
                },
                "messageDelay": {
                  "type": "number",
                  "example": 3600
                },
                "priority": {
                  "type": "string",
                  "example": "string"
                },
                "sender": {
                  "type": "string",
                  "example": "string"
                },
                "enclosed_file": {
                  "type": "string",
                  "example": "string"
                },
                "bindingKeys": {
                  "type": "string",
                  "example": "number[]"
                }
              },
              "required": [
                "title",
                "body",
                "messageTtl",
                "messageDelay",
                "priority",
                "sender",
                "enclosed_file",
                "bindingKeys"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned notif predefined message id",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                },
                "title": {
                  "type": "string",
                  "example": "string"
                },
                "body": {
                  "type": "string",
                  "example": "string"
                },
                "messageTtl": {
                  "type": "number",
                  "example": 3600
                },
                "messageDelay": {
                  "type": "number",
                  "example": 3600
                },
                "priority": {
                  "type": "string",
                  "example": "string"
                },
                "sender": {
                  "type": "string",
                  "example": "string"
                }
              },
              "required": [
                "id",
                "title",
                "body",
                "messageTtl",
                "messageDelay",
                "priority",
                "sender"
              ],
              "xml": {
                "name": "main"
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "NOTIF MESSAGES"
        ],
        "summary": "get all notif predefined messages ",
        "description": "",
        "responses": {
          "200": {
            "description": "Returned notif predefined messages",
            "schema": {
              "type": "array",
              "items": {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "number",
                    "example": 1
                  },
                  "title": {
                    "type": "string",
                    "example": "string"
                  },
                  "body": {
                    "type": "string",
                    "example": "string"
                  },
                  "messageTtl": {
                    "type": "number",
                    "example": 3600
                  },
                  "messageDelay": {
                    "type": "number",
                    "example": 3600
                  },
                  "priority": {
                    "type": "string",
                    "example": "string"
                  },
                  "sender": {
                    "type": "string",
                    "example": "string"
                  }
                },
                "required": [
                  "id",
                  "title",
                  "body",
                  "messageTtl",
                  "messageDelay",
                  "priority",
                  "sender"
                ]
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/notifMessages/{id}": {
      "put": {
        "tags": [
          "NOTIF MESSAGES"
        ],
        "summary": "update a notif predefined message",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "message",
            "in": "body",
            "description": "update notif predefined message",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "string"
                },
                "body": {
                  "type": "string",
                  "example": "string"
                },
                "messageTtl": {
                  "type": "number",
                  "example": 3600
                },
                "messageDelay": {
                  "type": "number",
                  "example": 3600
                },
                "priority": {
                  "type": "string",
                  "example": "string"
                },
                "sender": {
                  "type": "string",
                  "example": "string"
                }
              },
              "required": [
                "title",
                "body",
                "messageTtl",
                "messageDelay",
                "priority",
                "sender"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      },
      "delete": {
        "tags": [
          "NOTIF MESSAGES"
        ],
        "summary": "delete a notif predefined message",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "notif message ID."
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
    "/notifMessages/send": {
      "post": {
        "tags": [
          "NOTIF MESSAGES"
        ],
        "summary": "send a Notif Message",
        "description": "",
        "parameters": [
          {
            "name": "message",
            "in": "body",
            "description": "send notif message ",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "string"
                },
                "body": {
                  "type": "string",
                  "example": "string"
                },
                "messageTtl": {
                  "type": "number",
                  "example": 3600
                },
                "messageDelay": {
                  "type": "number",
                  "example": 0
                },
                "priority": {
                  "type": "string",
                  "example": "string"
                },
                "sender": {
                  "type": "string",
                  "example": "string"
                },
                "enclosed_file": {
                  "type": "string",
                  "example": "string"
                },
                "bindingKeys": {
                  "type": "array",
                  "example": [
                    1,
                    2
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              },
              "required": [
                "title",
                "body",
                "messageTtl",
                "messageDelay",
                "priority",
                "sender",
                "enclosed_file",
                "bindingKeys"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned notif message id",
            "schema": {
              "type": "array",
              "example": [
                true
              ],
              "items": {
                "type": "boolean"
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/notifMessages/sent/{id}": {
      "delete": {
        "tags": [
          "NOTIF MESSAGES"
        ],
        "summary": "delete a sent notif message",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "notif message ID."
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/notifMessages/delay": {
      "post": {
        "tags": [
          "NOTIF MESSAGES"
        ],
        "summary": "send a Notif Message",
        "description": "",
        "parameters": [
          {
            "name": "message",
            "in": "body",
            "description": "send notif message ",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "string"
                },
                "body": {
                  "type": "string",
                  "example": "string"
                },
                "messageTtl": {
                  "type": "number",
                  "example": 3600
                },
                "messageDelay": {
                  "type": "number",
                  "example": 3600
                },
                "priority": {
                  "type": "string",
                  "example": "string"
                },
                "sender": {
                  "type": "string",
                  "example": "string"
                },
                "enclosed_file": {
                  "type": "string",
                  "example": "string"
                },
                "bindingKeys": {
                  "type": "array",
                  "example": [
                    1,
                    2
                  ],
                  "items": {
                    "type": "number"
                  }
                }
              },
              "required": [
                "title",
                "body",
                "messageTtl",
                "messageDelay",
                "priority",
                "sender",
                "enclosed_file",
                "bindingKeys"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Returned notif message id",
            "schema": {
              "type": "object",
              "properties": {
                "id": {
                  "type": "number",
                  "example": 1
                }
              },
              "required": [
                "id"
              ],
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/notifMessages/delay/{historyId}": {
      "put": {
        "tags": [
          "NOTIF MESSAGES"
        ],
        "summary": "update a delayed notif message",
        "description": "",
        "parameters": [
          {
            "name": "historyId",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "type": "integer",
            "description": "notif message ID.",
            "required": true
          },
          {
            "name": "message",
            "in": "body",
            "description": "update delayed notif message",
            "schema": {
              "type": "object",
              "properties": {
                "title": {
                  "type": "string",
                  "example": "string"
                },
                "body": {
                  "type": "string",
                  "example": "string"
                },
                "messageTtl": {
                  "type": "number",
                  "example": 3600
                },
                "messageDelay": {
                  "type": "number",
                  "example": 3600
                },
                "priority": {
                  "type": "string",
                  "example": "string"
                },
                "sender": {
                  "type": "string",
                  "example": "string"
                },
                "enclosed_file": {
                  "type": "string",
                  "example": "string"
                }
              },
              "required": [
                "title",
                "body",
                "messageTtl",
                "messageDelay",
                "priority",
                "sender",
                "enclosed_file"
              ]
            }
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
    "/notifMessages/delay/{jobId}/{messageId}": {
      "delete": {
        "tags": [
          "NOTIF MESSAGES"
        ],
        "summary": "cancel a delayed notif message",
        "description": "",
        "parameters": [
          {
            "name": "jobId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "job ID."
          },
          {
            "name": "messageId",
            "in": "path",
            "required": true,
            "type": "integer",
            "description": "message ID."
          }
        ],
        "responses": {
          "default": {
            "description": ""
          }
        }
      }
    },
...
  }
}